04/04/24 - Modified CSS Colors
I modified several colors in CSS. The biggest change was changing the background from dark gray to midnight blue. The h1 and h2 selectors were modified so that green and red are displayed, respectively.  Previously, they were purple.

The selector ".cl_style__button_blue" name was changed to ".cl_style__button_black". Previously, the button was misnamed because the button was black, not blue. Some changes were made to the button's CSS appearance including an addition of an active rule.

Elements that were previously colored with rgb(139, 30, 247) - purple were replaced with the color rgb(77, 33, 199) - violet (except for the h1 and h2 selectors) .

The stock price output with selector "#id_form__calculator_control_panel :nth-child(8)" was moved 6 pixels to the left.

The input selector's background-color rule was made slightly darker. These style the input boxes to the calculator.