08/23/24 - Added Calculator Size Limitation
I added a new function in stock_calc.js called "verify_model()". In index.html, verify_model() is called after every call to recompute_model(). verify_model() checks every input and output of the calculator table and calculator control panel to see if a JavaScript number is larger than 9,999,999.99. If it is, the calculator is reset and transferred to HTML. A string indicating that the limit was exceeded will appear in the Predicted Stock Price output.

I implemented this for two reasons. The first was that the calculator display would overflow in HTML if values exceeded 9,999,999.99. The idea is that the calculator is supposed to be used with thousands or millions of units of currency, not the actual amount. Secondly, limiting a generally a good idea. The Predicted Stock Price output (with a large enough stock price) can extend across the entire screen.