08/30/24 - Bug Fix to 9,999,999.99 Limitation
The size limitation of 9,999,999.99 that was created for the Web version's calculator did not apply to negative numbers.

Prior to this commit, values could be below -9,999.999.99. With this bug fix, input or output values cannot be less than -9,999,999.99 for the Web version.

I also increased the size of <td> elements in the calculator table to 77px. This was done to allow enough space for strings with the same number of characters that  "-9,999,999.99" has. Previously, a large negative number like "-9,999,999.99" would cause the <td> elements to overflow.

The calculator slider was also moved to the right by 22px (11 max columns * (77 px - 75 px)). The previous size of <td> elements in the calculator table was 75px.