08/31/24 - Bug Fix to Excel Version and Web Text Change
I fixed a bug in the Excel version. The bankers_round() VBA function needs to be saved to a macro-enabled workbook. If one is not used, bankers_round() will not be available on the next load of the document. The Excel version is now a macro-enabled workbook.

Secondly, the string "(millions)" has been removed from the description of the Excel version in index.html. My reasoning is that the phrase "millions "already appears in the Excel document. Also, the user is free to use whatever denomination they want. Specifically mentioning to use millions may confuse people.