10/03/24 - Site Name Change to Stock Valuation Resource (CSS Changes and Banner Added)
I changed most of the colors, the layout, and the nav bar. A banner was added to the site. The old logo was removed. The site is no longer referred to as "ValueVille" and is instead referred to as "Stock Valuation Resource".

The look of the site has changed significantly. Compare the prior commit to this commit to see the changes.